+++ govc/CHANGELOG.md
@@ -1,5 +1,9 @@
 # changelog
 
+### unreleased (N/A)
+
+* Add basic NVME controller support
+
 ### 0.11.4 (2016-11-15)
 
 * Add role create, remove, update, ls and usage commands